Split out gsoap related parts in static library
The autogenerated source files by gsoap are fragile to CFLAGS changes.
Setting -DDEBUG, for example, results in an undefined reference error if
gsoap is not compiled as debug version.
Split out the gsoap related parts into a separate static library,
compile, and link with the same flags as the installed library.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>