Consider pysimplesoap array handling
The counterpart of libr4d, the r4dd server, uses pysimplesoap for handling
SOAP requests. Unfortunately the array handling had a longstanding bug.
Arrays of complex elements like:
<array>
  <list>
    <name>A</name>
    <port>1234</port
  </list>
  <list>
    <name>B</name>
    <port>5678</port>
  </list>
  ...
</array>

mapped to a SOAP response and WSDL like:
...
  <list>
    <name>A</name>
    <name>B</name>
    ...
    <port>1234</port
    <port>5678</port>
    ...
  </list>
...

This bug was partly fixed in a the current pysimplesoap version. Except
for WDSL the response is now correct.

Change the WDSL file according to the new response and change the code
to fit the new autogenerated code.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>